Moved MainMenu and GameHandler to seperate classes
